requirements.txt
asgiref==3.9.0
django==5.2.4
djangorestframework==3.16.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
django_filter==25.1